// CRKing7_ADMIN/src/types/sale.type.ts
export interface Sale {
  id: number;
  name: string;
  discount: number;
  isActive: number;
  startDate: string;
  endDate: string;
  modifiedDate: string;
  createdDate: string;
}
// const bb ={
//   "id": 2,
//         "name": "Chương trình test 2",
//         "discount": 30,
//         "isActive": 1,
//         "startDate": "01:57:00 09-08-2023",
//         "endDate": "07:00:00 10-08-2023",
//         "modifiedDate": "2023-08-08T19:10:32.000+07:00",
//         "createdDate": "2023-08-08T19:10:32.000+07:00"
// }